pyproject.toml
[project]
name = "piece-of-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "pydantic[email,timezone]>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "streamlit-code-editor>=0.1.22",
    "watchdog>=6.0.0",
]
